requirements.txt
# SPDX-FileCopyrightText: 2023 Adam Wojdyła <adam.wojdyla@outlook.com>
#
# SPDX-License-Identifier: MIT

pylint~=2.15.7
pycodestyle~=2.10.0
pycodestyle_magic~=0.5
opencv-python~=4.5.5.62
torch~=1.13.0
torchvision~=0.14.0
scikit-image~=0.19.3
scikit-learn~=1.1.3
numpy~=1.23.5
jupyter~=1.0.0
matplotlib~=3.6.2
seaborn~=0.12.1
tqdm~=4.64.1
mlflow==2.1.1
mlflow-skinny==2.1.1
azureml-mlflow==1.48.0
cloudpickle==2.2.0
black==22.12.0
azureml-core==1.48.0
hydra-core==1.3.2
omegaconf==2.3.0